refactor(security): isolate validation helpers
