+++ src/codex_plugin_scanner/guard/daemon/server.py
@@ -41,12 +41,8 @@ from ..approval_gate import (
     disable_totp,
     require_high_risk,
 )
-from ..approval_gate import (
-    input_from_mapping as approval_gate_input_from_mapping,
-)
-from ..approval_gate import (
-    public_config as approval_gate_public_config,
-)
+from ..approval_gate import input_from_mapping as approval_gate_input_from_mapping
+from ..approval_gate import public_config as approval_gate_public_config
 from ..approval_gate import (
     revoke_cooldown as revoke_approval_gate_cooldown,
 )
@@ -118,6 +114,7 @@ from ..desktop_notifications import (
 )
 from ..harness_disconnect_gate import require_harness_disconnect_gate
 from ..insights_share import publish_insights_share
+from ..json_transport import escape_json_for_html
 from ..local_dashboard_session import (
     DEFAULT_LOCAL_DASHBOARD_SESSION_TTL_SECONDS,
     LOCAL_DASHBOARD_SESSION_AUDIENCE,
@@ -451,24 +448,10 @@ def _codex_live_replay_authority(
 
 
 _PEER_DISCONNECT_ERRORS = (BrokenPipeError, ConnectionResetError, ConnectionAbortedError)
-_JSON_RESPONSE_TRANSLATION = str.maketrans(
-    {
-        "&": "\\u0026",
-        "<": "\\u003c",
-        ">": "\\u003e",
-    }
-)
-
-
-def _json_response_bytes(payload: dict[str, Any]) -> bytes:
-    """Serialize JSON without leaving HTML-significant characters on the wire."""
-
-    return json.dumps(payload).translate(_JSON_RESPONSE_TRANSLATION).encode("utf-8")
 
 
 class _GuardDaemonHTTPServer(BoundedThreadingHTTPServer):
     request_queue_size = _MAX_CONCURRENT_DAEMON_CONNECTIONS
-
     store: GuardStore
     runtime: GuardSurfaceRuntime
     auth_token: str
@@ -7674,15 +7657,14 @@ class _GuardDaemonHandler(BaseHTTPRequestHandler):
         status: int = 200,
         extra_headers: dict[str, str] | None = None,
     ) -> None:
-        body = _json_response_bytes(payload)
-        headers = dict(extra_headers or {})
+        body = escape_json_for_html(json.dumps(payload).encode("utf-8"))
+        headers = {**dict(extra_headers or {}), "X-Content-Type-Options": "nosniff"}
         cors_headers = self._cors_headers_for_request(allow_methods="GET, POST, OPTIONS")
         if cors_headers is not None:
             headers = {**cors_headers, **headers}
         try:
             self.send_response(status)
             self.send_header("Content-Type", "application/json; charset=utf-8")
-            self.send_header("X-Content-Type-Options", "nosniff")
             self.send_header("Content-Length", str(len(body)))
             for key, value in self._validated_headers(headers).items():
                 self.send_header(key, value)
@@ -7717,6 +7699,7 @@ class _GuardDaemonHandler(BaseHTTPRequestHandler):
             "Location",
             "Pragma",
             "Vary",
+            "X-Content-Type-Options",
         }
         validated: dict[str, str] = {}
         for key, value in (extra_headers or {}).items():
